api/requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.1
pg8000==1.30.5
